mm_util: change lengths to size_t

+++ mm_manager.h
@@ -987,12 +987,12 @@ extern int hangup_modem(int fd);
 
 /* mm_util */
 extern unsigned crc16(unsigned crc, uint8_t *buf, size_t len);
-extern void dump_hex(uint8_t *data, int len);
-extern char *phone_num_to_string(char *string_buf, int string_len, uint8_t* num_buf, int num_buf_len);
+extern void dump_hex(uint8_t *data, size_t len);
+extern char *phone_num_to_string(char *string_buf, size_t string_len, uint8_t* num_buf, size_t num_buf_len);
 extern uint8_t string_to_bcd_a(char* number_string, uint8_t* buffer, uint8_t buff_len);
-extern char *callscrn_num_to_string(char *string_buf, int string_buf_len, uint8_t* num_buf, int num_buf_len);
-extern char *call_type_to_string(uint8_t call_type, char *string_buf, int string_buf_len);
-extern char *timestamp_to_string(uint8_t *timestamp, char *string_buf, int string_buf_len);
+extern char *callscrn_num_to_string(char *string_buf, size_t string_buf_len, uint8_t* num_buf, size_t num_buf_len);
+extern char *call_type_to_string(uint8_t call_type, char *string_buf, size_t string_buf_len);
+extern char *timestamp_to_string(uint8_t *timestamp, char *string_buf, size_t string_buf_len);
 extern void print_bits(uint8_t bits, char *str_array[]);
 extern char *table_to_string(uint8_t table);
